Remove obsolete blender file, rename shaders
Import of glTF meshes into blender is so easy that maintaining
an extra .blend file is obsolete. Therefore .blend file and Blender
folders are deleted here. Removed the Godot subfolder as now Godot
is the only tool used in this code project.
Renamed shaders from .shader to .gdshader as only the latest will
be supported in Godot 4.