Simplify project adaptions for own games
Game icon and start screen background are moved to the level-assets
folder which (together with level) would be exchanged in an own
game. Create a game title from project name by replacing '-', '_'
and ' ' with line breaks. Add a project property for the website
called by the 'Game infos' button. The button is deactivated in
case of empty property string.